[WIP] - Fix wrong state reset
